+++ src/oic/utils/client_management.py
@@ -9,7 +9,7 @@ import argparse
 import sys
 from future.backports.urllib.parse import urlparse
 from future.backports.urllib.parse import splitquery
-from future.backports.urllib.parse import parse_qs
+from future.moves.urllib.parse import parse_qs
 
 from oic.oic.provider import secret
 from oic import rndstr